[z2051] Apply CREATURE_FLAG_EXTRA_NO_TALKTO_CREDIT to all creatures
The flag was introduced in [s1442]. It is assumed that by now all quests that require some talk-to credit are handled by different means.
If you catch quests that don't work yet, remove this flag for this npc and report to your DB provider.

(based on commit 629803e)

(based on commit [s1665] - 53fa0af)